feat: add support for snapshot manager backing the role
Snapshot Manager is available on Centos 10, RHEL 9 and RHEL 10.

Update to be compatiable with legacy code that is already deployed.

Note that Snapshot Manager has minimum sizes for any
snapshot of an origin.  If the minimum size is used, an extend may
not expand the size if the additional space is still below the
minimum size.  Currently the minimum size for snapm is 512 MiB.
Future versions will decrease to 64 MiB.  Some adjustment to the
extend tests were made in this change to deal with failures of
verify expecting expand to return true when the requested size was still
below the minimum.
