+++ .ostree/packages-runtime-CentOS-10.txt
@@ -1,1 +1,3 @@
 util-linux-core
+snapm
+boom-boot
