+++ module_utils/snapshot_lsr/consts.py
@@ -1,0 +1,55 @@
+class SnapshotCommand:
+    SNAPSHOT = "snapshot"
+    CHECK = "check"
+    REMOVE = "remove"
+    REVERT = "revert"
+    EXTEND = "extend"
+    LIST = "list"
+    MOUNT = "mount"
+    UMOUNT = "umount"
+    INVALID = "invalid"
+
+
+class SnapshotStatus:
+    SNAPSHOT_OK = 0
+    ERROR_INSUFFICIENT_SPACE = 1
+    ERROR_ALREADY_DONE = 2
+    ERROR_SNAPSHOT_FAILED = 3
+    ERROR_REMOVE_FAILED = 4
+    ERROR_REMOVE_FAILED_NOT_SNAPSHOT = 5
+    ERROR_REMOVE_FAILED_INUSE = 6
+    ERROR_REMOVE_FAILED_NOT_FOUND = 7
+    ERROR_LVS_FAILED = 8
+    ERROR_NAME_TOO_LONG = 9
+    ERROR_ALREADY_EXISTS = 10
+    ERROR_NAME_CONFLICT = 11
+    ERROR_VG_NOTFOUND = 12
+    ERROR_LV_NOTFOUND = 13
+    ERROR_VERIFY_NOTSNAPSHOT = 14
+    ERROR_VERIFY_COMMAND_FAILED = 15
+    ERROR_VERIFY_NOT_FOUND = 16
+    ERROR_CMD_INVALID = 17
+    ERROR_VERIFY_REMOVE_FAILED = 18
+    ERROR_VERIFY_REMOVE_SOURCE_SNAPSHOT = 19
+    ERROR_SNAPSET_SOURCE_DOES_NOT_EXIST = 20
+    ERROR_SNAPSET_CHECK_STATUS_FAILED = 21
+    ERROR_SNAPSET_INSUFFICIENT_SPACE = 22
+    ERROR_JSON_PARSER_ERROR = 23
+    ERROR_INVALID_PERCENT_REQUESTED = 24
+    ERROR_UNKNOWN_FAILURE = 25
+    ERROR_REVERT_FAILED = 26
+    ERROR_EXTEND_NOT_SNAPSHOT = 27
+    ERROR_EXTEND_NOT_FOUND = 28
+    ERROR_EXTEND_FAILED = 29
+    ERROR_EXTEND_VERIFY_FAILED = 29
+    ERROR_UMOUNT_FAILED = 30
+    ERROR_MOUNT_FAILED = 31
+    ERROR_MOUNT_POINT_NOT_EXISTING = 32
+    ERROR_MOUNT_NOT_BLOCKDEV = 33
+    ERROR_MOUNT_INVALID_PARAMS = 34
+    ERROR_MOUNT_POINT_ALREADY_MOUNTED = 35
+    ERROR_MOUNT_VERIFY_FAILED = 36
+    ERROR_UMOUNT_VERIFY_FAILED = 37
+    ERROR_UMOUNT_NOT_MOUNTED = 38
+    ERROR_CREATE_FAILED = 39
+    ERROR_SNAPM_INTERNAL_ERROR = 40
