+++ tests/roles/linux-system-roles.snapshot/module_utils
@@ -1,0 +1,1 @@
+../../../module_utils
