+++ tests/tests_no_space_fail.yml
@@ -21,7 +21,7 @@
         - name: Test failure of creating snapshot
           include_tasks: verify-role-failed.yml
           vars:
-            __snapshot_failed_regex: insufficient space for snapshots
+            __snapshot_failed_regex: "insufficient space for snapshots*|Insufficient free space for snapshot*"
             __snapshot_failed_msg: Role did not fail with no space error
             __snapshot_failed_params:
               snapshot_lvm_percent_space_required: 15
