+++ tests/tests_set_extend.yml
@@ -1,7 +1,9 @@
 ---
-- name: Revert snapshots of logical volumes across different volume groups
+- name: Extend snapshots of logical volumes across different volume groups
   hosts: all
   vars:
+    # only use vgs matching this pattern
+    snapshot_lvm_vg_include: "^test_"
     test_disk_min_size: "1g"
     test_disk_count: 10
     test_storage_pools:
@@ -11,14 +13,14 @@
           - name: lv1
             size: "15%"
           - name: lv2
-            size: "50%"
+            size: "15%"
       - name: test_vg2
         disks: "{{ range(3, 6) | map('extract', unused_disks) | list }}"
         volumes:
           - name: lv3
-            size: "10%"
+            size: "15%"
           - name: lv4
-            size: "20%"
+            size: "15%"
       - name: test_vg3
         disks: "{{ range(6, 10) | map('extract', unused_disks) | list }}"
         volumes:
@@ -26,48 +28,44 @@
             size: "30%"
           - name: lv6
             size: "25%"
-          - name: lv7
-            size: "10%"
-          - name: lv8
-            size: "10%"
     snapshot_test_set:
       name: snapset1
       volumes:
         - name: snapshot VG1 LV1
           vg: test_vg1
           lv: lv1
-          percent_space_required: 15
+          percent_space_required: 50
         - name: snapshot VG2 LV3
           vg: test_vg2
           lv: lv3
-          percent_space_required: 15
+          percent_space_required: 50
         - name: snapshot VG2 LV4
           vg: test_vg2
           lv: lv4
-          percent_space_required: 15
-        - name: snapshot VG3 LV7
+          percent_space_required: 50
+        - name: snapshot VG3 LV6
           vg: test_vg3
-          lv: lv7
-          percent_space_required: 15
+          lv: lv6
+          percent_space_required: 50
     snapshot_extend_set:
       name: snapset1
       volumes:
         - name: snapshot VG1 LV1
           vg: test_vg1
           lv: lv1
-          percent_space_required: 30
+          percent_space_required: 60
         - name: snapshot VG2 LV3
           vg: test_vg2
           lv: lv3
-          percent_space_required: 30
+          percent_space_required: 60
         - name: snapshot VG2 LV4
           vg: test_vg2
           lv: lv4
-          percent_space_required: 30
-        - name: snapshot VG3 LV7
+          percent_space_required: 60
+        - name: snapshot VG3 LV6
           vg: test_vg3
-          lv: lv7
-          percent_space_required: 30
+          lv: lv6
+          percent_space_required: 60
   tasks:
     - name: Run tests
       block:
@@ -96,6 +94,10 @@
             snapshot_lvm_action: extend
             snapshot_lvm_set: "{{ snapshot_extend_set }}"
 
+        # Both LVM and Snapshot Manager have minimum sizes for snapshots.
+        # extend may return changed=false if the original request was
+        # smaller than the minimum snapshot size and the new percentage
+        # is also smaller than the original.
         - name: Assert changes for extend
           assert:
             that: snapshot_cmd["changed"]
