+++ tests/tests_set_extend_verify_fail.yml
@@ -16,14 +16,14 @@
         - name: snapshot VG1 LV1
           vg: test_vg1
           lv: lv1
-          percent_space_required: 20
+          percent_space_required: 10
     snapshot_test_verify_set:
       name: snapset1
       volumes:
         - name: snapshot VG1 LV1
           vg: test_vg1
           lv: lv1
-          percent_space_required: 50
+          percent_space_required: 90
   tasks:
     - name: Run tests
       block:
