+++ tests/tests_set_mount_verify_fail.yml
@@ -1,5 +1,5 @@
 ---
-- name: Verify the extend verify command fails when space too low
+- name: Test mount fails when no snapset exists
   hosts: all
   vars:
     test_disk_min_size: "1g"
@@ -9,35 +9,68 @@
         disks: "{{ range(0, 3) | map('extract', unused_disks) | list }}"
         volumes:
           - name: lv1
+            size: "15%"
+          - name: lv2
             size: "50%"
+      - name: test_vg2
+        disks: "{{ range(3, 6) | map('extract', unused_disks) | list }}"
+        volumes:
+          - name: lv3
+            size: "10%"
+          - name: lv4
+            size: "20%"
+      - name: test_vg3
+        disks: "{{ range(6, 10) | map('extract', unused_disks) | list }}"
+        volumes:
+          - name: lv5
+            size: "30%"
+          - name: lv6
+            size: "25%"
+          - name: lv7
+            size: "10%"
+          - name: lv8
+            size: "10%"
     snapshot_test_set:
       name: snapset1
       volumes:
         - name: snapshot VG1 LV1
           vg: test_vg1
           lv: lv1
-          percent_space_required: 20
-    snapshot_test_verify_set:
-      name: snapset1
-      volumes:
-        - name: snapshot VG1 LV1
-          vg: test_vg1
-          lv: lv1
-          percent_space_required: 50
+          mountpoint: "{{ test_mnt_parent ~ '/lv1_mp' }}"
+          percent_space_required: 15
+          mountpoint_create: false
+        - name: snapshot VG2 LV3
+          vg: test_vg2
+          lv: lv3
+          mountpoint: "{{ test_mnt_parent ~ '/lv3_mp' }}"
+          percent_space_required: 15
+          mountpoint_create: false
+        - name: snapshot VG2 LV4
+          vg: test_vg2
+          lv: lv4
+          mountpoint: "{{ test_mnt_parent ~ '/lv4_mp' }}"
+          percent_space_required: 15
+          mountpoint_create: false
+        - name: snapshot VG3 LV7
+          vg: test_vg3
+          lv: lv7
+          mountpoint: "{{ test_mnt_parent ~ '/lv7_mp' }}"
+          percent_space_required: 15
+          mountpoint_create: false
   tasks:
     - name: Run tests
       block:
         - name: Setup
           include_tasks: tasks/setup.yml
 
-        - name: Run the snapshot role to create snapshot set of LVs
+        - name: Run the snapshot role to create a snapshot set of LVs
           include_role:
             name: linux-system-roles.snapshot
           vars:
             snapshot_lvm_action: snapshot
             snapshot_lvm_set: "{{ snapshot_test_set }}"
 
-        - name: Run the snapshot role to verify the set of snapshots for the LVs
+        - name: Verify the set of snapshots for the LVs
           include_role:
             name: linux-system-roles.snapshot
           vars:
@@ -45,16 +78,43 @@
             snapshot_lvm_set: "{{ snapshot_test_set }}"
             snapshot_lvm_verify_only: true
 
-        - name: Test failure extend verify
+        - name: Mount the set
+          include_role:
+            name: linux-system-roles.snapshot
+          vars:
+            snapshot_lvm_action: mount
+            snapshot_lvm_mountpoint_create: true
+            snapshot_lvm_set: "{{ snapshot_test_set }}"
+
+        - name: Assert changes for mount
+          assert:
+            that: snapshot_cmd["changed"]
+
+        - name: Verify the mount is done
+          include_role:
+            name: linux-system-roles.snapshot
+          vars:
+            snapshot_lvm_verify_only: true
+            snapshot_lvm_action: mount
+            snapshot_lvm_set: "{{ snapshot_test_set }}"
+
+        - name: Umount the set
+          include_role:
+            name: linux-system-roles.snapshot
+          vars:
+            snapshot_lvm_action: umount
+            snapshot_lvm_set: "{{ snapshot_test_set }}"
+
+        - name: Test mount verify failure
           include_tasks: verify-role-failed.yml
           vars:
             __snapshot_failed_regex:
-              "verify failed due to insufficient space for:*"
+              "blockdev not mounted on any mountpoint:*"
             __snapshot_failed_msg: Role did not fail with extend verify
             __snapshot_failed_params:
-              snapshot_lvm_action: extend
+              snapshot_lvm_action: mount
               snapshot_lvm_verify_only: true
-              __snapshot_lvm_set: "{{ snapshot_test_verify_set }}"
+              __snapshot_lvm_set: "{{ snapshot_test_set }}"
             __snapshot_failed_changed: false
 
       always:
