+++ vars/CentOS_10.yml
@@ -1,0 +1,1 @@
+RedHat_8.yml
