+++ vars/RedHat_7.yml
@@ -1,4 +1,3 @@
 # SPDX-License-Identifier: MIT
 ---
-__snapshot_packages: ["lvm2", "util-linux"]
 __snapshot_python: /usr/bin/python
