+++ vars/main.yml
@@ -1,6 +1,18 @@
 # SPDX-License-Identifier: MIT
 ---
-__snapshot_packages: ["lvm2", "util-linux-core"]
+__snapshot_base_packages_7_8: ["lvm2", "util-linux"]
+__snapshot_base_packages_other: ["lvm2", "util-linux-core"]
+__snapshot_base_packages: "{{ __snapshot_base_packages_7_8
+  if (__snapshot_is_rh_distro and
+      ansible_distribution_version is version('9', '<'))
+  else __snapshot_base_packages_other }}"
+__snapshot_extra_packages: ["boom-boot", "snapm"]
+__snapshot_packages: "{{ __snapshot_base_packages + __snapshot_extra_packages
+  if ((__snapshot_is_rh_distro and
+       ansible_distribution_version is version('9.6', '>=')) or
+      (ansible_distribution == 'Fedora' and
+       ansible_distribution_version is version('41', '>=')))
+  else __snapshot_base_packages }}"
 __snapshot_services: []
 # python to use on managed nodes for snapshot.py
 __snapshot_python: /usr/bin/python3
